// ic-elp/src/app/domain/enums/request-enums.ts
export enum RequestType {
  personalLeave = 'Personal Leave',
  onDemandLeave = 'On demand leave',
  occassionalLeave = 'Occassional leave',
  excusedPaidAbsence = 'Excused paid absence',
  excusedUnpaidAbsence = 'Excused unpaid absence',
  unpaidLeave = 'Unpaid leave',
  childCare = 'Child care',
  maternityLeave = 'Maternity leave',
  parentalLeave = 'Parental leave',
  paternityLeave = 'Paternity leave',
  unpaidChildcareLeave = 'Unpaid childcare leave'
}

export enum Occassion {
  employeesWedding = 'Employee’s wedding',
  emploeesChildWedding = 'Emploee’s child wedding',
  death = 'Death',
  childBirth = 'Child birth'
}

export enum DeathOf {
  spouse = 'Spouse',
  child = 'Child',
  mother = 'Mother',
  father = 'Father',
  fatherInLaw = 'Father-in-law',
  motherInLaw = 'Mother-in-law',
  grandfather = 'Grandfather',
  grandmother = 'Grandmother',
  stepfather = 'Stepfather',
  stepmother = 'Stepmother',
  sister = 'Sister',
  brother = 'Brother',
  other = 'Other dependent or person being under employee’s care'
}